move card to nx libs
