+++ libs/ui/card/README.md
@@ -1,0 +1,7 @@
+# card
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test card` to execute the unit tests.
