+++ libs/ui/card/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/card/card.component';
