--- src/app/presentation/modules/card/index.ts
@@ -1,2 +1,0 @@
-
-export * from './card.component';
