+++ tsconfig.app.json
@@ -13,6 +13,7 @@
       "@data/*": ["src/app/data/*"],
       "@ui/toaster": ["libs/ui/toaster/src/index.ts"],
       "@ui/listbox": ["libs/ui/listbox/src/index.ts"],
+      "@ui/card": ["libs/ui/card/src/index.ts"],
       "@ui/*": ["src/app/presentation/*"]
 
     }
