+++ tsconfig.base.json
@@ -17,8 +17,9 @@
       "@data/*": ["src/app/data/*"],
       "@shared/models": ["libs/shared/models/src/index.ts"],
       "@state": ["libs/state/src/index.ts"],
-      "@ui/toaster": ["libs/ui/toaster/src/index.ts"],
       "@ui/listbox": ["libs/ui/listbox/src/index.ts"],
+      "@ui/toaster": ["libs/ui/toaster/src/index.ts"],
+      "@ui/card": ["libs/ui/card/src/index.ts"],
       "@ui/*": ["src/app/presentation/*"]
     }
   }
